requirements.txt
python-dotenv~=0.15.0
asyncpg~=0.22.0
aiogram-dev2==2.9.2